+++ tests/test_manifold_embedding_consistency.py
@@ -67,15 +67,17 @@ def setup_and_test_multiple_pdfs(pdf_def, layer_def):
 
     seed_everything(0)
     normal_pdf=f.pdf(pdf_def, layer_def)
+    normal_pdf.double()
 
     seed_everything(0)
     embedded_pdf=f.pdf(pdf_def, layer_def)
     embedded_pdf.set_use_embedding_parameters_flag(True)
+    embedded_pdf.double()
 
     seed_everything(0)
     intrinsic_pdf=f.pdf(pdf_def, layer_def)
     intrinsic_pdf.set_use_embedding_parameters_flag(False)
-
+    intrinsic_pdf.double()
 
     for sub_pdf_index in [-1]+list(range(len(normal_pdf.layer_list))):
         print("sub pdf ", sub_pdf_index)
